% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrowButtonGroupInput.R
\name{arrowButtonGroupInput}
\alias{arrowButtonGroupInput}
\title{Generates arrow buttons}
\usage{
arrowButtonGroupInput(inputId, type = "h", origin = c(0L, 0L))
}
\description{
Generates arrow buttons
}
