% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drilldownDTModule.R
\name{drilldownDTModule}
\alias{drilldownDTModule}
\title{server for drillDTModalModule}
\usage{
drilldownDTModule(input, output, session, longdf, type = c("box",
  "modal"), level_cols, value_col, agg_fn = sum, pivot_col = NULL,
  pivot_fill = NA, transpose = FALSE, options = NULL)
}
\description{
server for drillDTModalModule
}
