% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drilldownDTModule.R
\name{fgap_df}
\alias{fgap_df}
\title{filter then group then aggregate then pivot a data frame}
\usage{
fgap_df(df, filters = NULL, group_col, value_col, agg_fn = sum,
  pivot_col = NULL, pivot_fill = NA, transpose = FALSE)
}
\description{
filter then group then aggregate then pivot a data frame
}
