% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_number}
\alias{format_number}
\title{using formattable functions to format a vector given its type}
\usage{
format_number(x, type, digits = 0, symbol = "$",
  remove_infinite = TRUE)
}
\arguments{
\item{x}{a numeric vector}

\item{type}{type for the vector, valid values are "currency", "percent", 
"comma" and "accounting"}

\item{digits}{digits for the number}

\item{symbol, }{currency symbol. only works when type is currency.}

\item{remove_infinite}{default to TRUE. Conver NA,NaN, Inf ... to ""}
}
\value{
a character vector with formatted numbers
}
\description{
using formattable functions to format a vector given its type
}
