% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSDTModule.R
\name{FSDTModule}
\alias{FSDTModule}
\title{Server of FSDT Module}
\usage{
FSDTModule(input, output, session, id, df, extensions = c("FixedHeader",
  "FixedColumns"), dt_options = "default")
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny outpu}

\item{session}{shiny session}

\item{id}{id for the module}

\item{df}{a data frame, preferably generated with \link{createFSTable}}

\item{extensions}{a character vector, default to 
\code{c("FixedHeader",'FixedColumns')}. value is passed to \link{DT::datatable}}

\item{dt_options}{"default" or a list. value is passed to \code{options} 
argument in \link{DT::datatable}}
}
\value{
generates a datatable in DT package with predefined or custom attributes
}
\description{
Server of FSDT Module
}
\seealso{
\link{createFSTable}, \link{FSDTModuleUI}
}
