% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSDTModule.R
\name{FSDTSampleData}
\alias{FSDTSampleData}
\title{return a list that contains all sample data required to make a FSDT}
\usage{
FSDTSampleData()
}
\value{
a list which contains all data required to make a FSDT
}
\description{
FSDT is the abbreviation of Financial Statement Data Table.
}
\seealso{
\code{\link{createFSTable}}
}
