% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobileDevice.R
\name{isMobile}
\alias{isMobile}
\title{inserts JavaScript to detect if user is using Mobile Device}
\usage{
isMobile()
}
\description{
inserts JavaScript to detect if user is using Mobile Device
}
\details{
use \code{input$isMobile} to receive if user is using Mobile Device. the value
is either \code{TRUE} or \code{FALSE}.
}
\examples{
if (interactive()){
shinyApp(
    ui <- fluidPage(
        isMobile()
    ),
    server <- function(input, output, session){
        isMobile <- eventReactive(input$isMobile,input$isMobile)
        observe(print(res()))
    }   
    )
}
}
