% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairedDTPlot.R
\name{pairedDTPlot}
\alias{pairedDTPlot}
\title{pairedDTplot Server}
\usage{
pairedDTPlot(input, output, session, df, width_dt = 8, width_plot = 4,
  title = "", value_type = "general", direction = "v",
  h_ignore_columns = NULL, height = "300px", server = FALSE)
}
\description{
pairedDTplot Server
}
