% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairedDTPlotModule.R
\name{pairedDTPlotModule}
\alias{pairedDTPlotModule}
\title{pairedDTplot Server}
\usage{
pairedDTPlotModule(input, output, session, df, width_dt = 8,
  width_plot = 4, title = "", value_type = "general",
  direction = "v", h_ignore_columns = NULL, h_reverse = FALSE,
  height = "300px", server = FALSE)
}
\description{
pairedDTplot Server
}
