% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectArrowInputModule.R
\name{selectArrowInputModule}
\alias{selectArrowInputModule}
\title{Server for selectArrowInputModule}
\usage{
selectArrowInputModule(input, output, session, inputId, label, choices,
  selected = NULL, selectize = TRUE, width = NULL, size = NULL)
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}

\item{InputId, }{label, ..., size are all arguments for selectInput}
}
\value{
Renders UI and returns selected value as a reactive value
}
\description{
Server for selectArrowInputModule
}
