% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectArrowInputModule.R
\name{selectArrowInputModuleUI}
\alias{selectArrowInputModuleUI}
\title{UI placeholder for selectArrowInputModule}
\usage{
selectArrowInputModuleUI(inputId)
}
\arguments{
\item{inputId}{inputId}
}
\value{
UI component
}
\description{
UI placeholder for selectArrowInputModule
}
