% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_meta.R
\name{shinyInit}
\alias{shinyInit}
\title{run this function to quickly initialize a shiny app folder
 @description 
 By running this function, it will create following files in given path:
 1. app.R if mode is "single"; ui.R, server.R, global.R if mode is "multiple"
 2. Folders: www/, www/css, www/js, www/img
 
 Behavior: this function won't overwrite existing files
 
 @param path path to folder
 @param mode what files to create, see more in description
 @param overwrite whether to overwrite existing files
 
 @return NULL, creates folders and files instead
 @export}
\usage{
shinyInit(path, mode = c("single", "multiple"))
}
\description{
run this function to quickly initialize a shiny app folder
 @description 
 By running this function, it will create following files in given path:
 1. app.R if mode is "single"; ui.R, server.R, global.R if mode is "multiple"
 2. Folders: www/, www/css, www/js, www/img
 
 Behavior: this function won't overwrite existing files
 
 @param path path to folder
 @param mode what files to create, see more in description
 @param overwrite whether to overwrite existing files
 
 @return NULL, creates folders and files instead
 @export
}
