% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrowButtonGroupInput.R
\name{arrowButtonGroupInput}
\alias{arrowButtonGroupInput}
\title{Generates arrow buttons}
\usage{
arrowButtonGroupInput(inputId, type = "h", size = "100\%",
  color = "rgb(0,0,0)", colorClick = "rgb(128,128,128)")
}
\description{
Generates arrow buttons
}
